Minimal playground implementation for demo and testing purposes
