+++ .gitignore
@@ -33,3 +33,6 @@ out/
 # Coverage
 .coverage
 htmlcov/
+
+run_server.sh
+run_client.sh
