+++ google/cloud/security/iam/explain/explainer.py
@@ -1,4 +1,4 @@
-import dao
+from google.cloud.security.iam import dao
 from importer import importer
 
 class Explainer():
