+++ google/cloud/security/iam/explain/utils.py
@@ -1,6 +1,1 @@
 
-import os
-import binascii
-
-def generateModelHandle():
-    return binascii.hexlify(os.urandom(16))
