+++ google/cloud/security/iam/playground/playground.proto
@@ -7,6 +7,93 @@ service Playground {
   rpc SetIamPolicy(SetIamPolicyRequest) returns (SetIamPolicyReply) {}
   rpc GetIamPolicy(GetIamPolicyRequest) returns (GetIamPolicyReply) {}
   rpc CheckIamPolicy(CheckIamPolicyRequest) returns (CheckIamPolicyReply) {}
+
+  rpc AddGroupMember(AddGroupMemberRequest) returns (AddGroupMemberReply) {}
+  rpc DelGroupMember(DelGroupMemberRequest) returns (DelGroupMemberReply) {}
+  rpc ListGroupMembers(ListGroupMembersRequest) returns (ListGroupMembersReply) {}
+
+  rpc AddResource(AddResourceRequest) returns (AddResourceReply) {}
+  rpc DelResource(DelResourceRequest) returns (DelResourceReply) {}
+  rpc ListResources(ListResourcesRequest) returns (ListResourcesReply) {}
+
+  rpc AddRole(AddRoleRequest) returns (AddRoleReply) {}
+  rpc DelRole(DelRoleRequest) returns (DelRoleReply) {}
+  rpc ListRoles(ListRolesRequest) returns (ListRolesReply) {}
+}
+
+message AddRoleRequest {
+  string role_name = 1;
+  repeated string permissions = 2;
+}
+
+message AddRoleReply {
+}
+
+message DelRoleRequest {
+  string role_name = 1;
+}
+
+message DelRoleReply {
+}
+
+message ListRolesRequest {
+  string prefix = 1;
+}
+
+message ListRolesReply {
+  repeated string role_names = 1;
+}
+
+message ListGroupMembersRequest {
+  string prefix = 1;
+}
+
+message ListGroupMembersReply {
+  repeated string member_names = 1;
+}
+
+message ListResourcesRequest {
+  string prefix = 1;
+}
+
+message ListResourcesReply {
+  repeated string full_resource_names = 1;
+}
+
+message AddGroupMemberRequest {
+  string member_name = 1;
+  string member_type = 2;
+  repeated string parent_names = 3;
+}
+
+message AddGroupMemberReply {
+}
+
+message DelGroupMemberRequest {
+  string member_name = 1;
+  string parent_name = 2;
+  bool only_delete_relationship = 3;
+}
+
+message DelGroupMemberReply {
+}
+
+message AddResourceRequest {
+  string full_resource_name = 1;
+  string resource_type = 2;
+  string parent_full_resource_name = 3;
+  bool no_require_parent = 4;
+}
+
+message AddResourceReply {
+}
+
+message DelResourceRequest {
+  string full_resource_name = 1;
+}
+
+message DelResourceReply {
+
 }
 
 message PingRequest {
@@ -27,10 +114,9 @@ message Policy {
 }
 
 message CheckIamPolicyRequest {
-	string model = 1;
-	string resource = 2;
-	string permission = 3;
-	string identity = 4;
+	string resource = 1;
+	string permission = 2;
+	string identity = 3;
 }
 
 message CheckIamPolicyReply {
@@ -38,10 +124,9 @@ message CheckIamPolicyReply {
 }
 
 message SetIamPolicyRequest {
-	string model = 1;
-	string resource = 2;
-	Policy policy = 3;
-	string etag = 4;
+	string resource = 1;
+	Policy policy = 2;
+	string etag = 3;
 }
 
 message SetIamPolicyReply {
@@ -49,8 +134,7 @@ message SetIamPolicyReply {
 }
 
 message GetIamPolicyRequest {
-	string model = 1;
-	string resource = 2;
+	string resource = 1;
 }
 
 message GetIamPolicyReply {
